qRevert "update readme"
This reverts commit ec46a42f5ce0dcabf73925492a167122df7f60d6, reversing
changes made to 275eea718a90bba280517ba5628938e1dffac005.
