+++ .gitignore
@@ -1,5 +1,3 @@
 *.elf
 *.img
 *.o
-initramfs.cpio
-*.out
