+++ README.md
@@ -7,48 +7,28 @@
 | `A091514`  | `kishow01`  | `林昌德` | kishow01@gmail.com         |
 
 ## How to build
-### bootloader
-```
-cd bootloader
-make
-```
-
-### kernel_sender
-```
-cd kernel_sender
-make
-```
-
-### cpio kernel
-```
-cd cpio_kernel
-make
-```
+``make``
 
 ## How to run
-### bootloader
-```
-cd bootloader
-make run
-```
-
-### cpio kernel
-```
-cd cpio_kernel
-make run
-```
+``make run``
 
 ## How to burn it into pi3
-Move bootloader.img, initramfs.cpio and config.txt into sd card
+download firmware from https://github.com/raspberrypi/firmware/tree/master/boot.
+
++ necessary
+    + bootcode.bin
+        + bootloader, loaded by the SoC on boot. 
+        + setup and loads one of the start*.elf files.
+    + fixup.dat
+        + linker files and are matched pairs with the start*.elf files
+    + start.elf
+        + binary blobs (firmware) that are loaded on to the VideoCore in the SoC, which then take over the boot process. start.elf is the basic firmware.
+
+Finally, put the firmware and your kernel image into the FAT partition.
 
 ## Architecture
 
-+ bootloader
-    + the uart bootloader in sd card
-+ kernel_sender
-    + send kernel through uart from host to rpi3
-+ kernel
-    + kernel that support cpio
+**WIP**
 
 ## Directory structure
