--- bootloader/type.h
@@ -1,10 +1,0 @@
-#ifndef _LCD_TYPE_H_
-#define _LCD_TYPE_H_
-
-typedef int int32_t;
-typedef long long int int64_t;
-
-typedef unsigned int uint32_t;
-typedef unsigned long long int uint64_t;
-
-#endif
