--- config.txt
@@ -1,4 +1,0 @@
-kernel_address=0x60000
-kernel=bootloader.img
-arm_64bit=1
-initramfs initramfs.cpio 0x8000000
