+++ main.c
@@ -8,5 +8,5 @@ void main() {
 
     while(1) {
         shell_start();
-    }
+    } 
 }
