--- rootfs/test1.txt
@@ -1,1 +1,0 @@
-test1
