--- rootfs/whereisflag/flag.txt
@@ -1,1 +1,0 @@
-flag{OSC_LAB2}
