--- script/cpio_gen.sh
@@ -1,3 +1,0 @@
-cd ../rootfs
-find . | cpio -o -H newc > ../initramfs.cpio
-cd ..
