+++ shell.h
@@ -1,0 +1,11 @@
+#ifndef _LCD_SHELL_H_
+#define _LCD_SHELL_H_
+
+void shell_welcome_message();
+void shell_start();
+
+void reboot();
+
+void command(char *cmd);
+
+#endif
